Imported Upstream version 1.2.99~20120228~SE~cb0b7f6~SYSYNC~5e0f652

+++ src/synthesis/src/sysync/stdlogicds.cpp
@@ -402,7 +402,7 @@ localstatus TStdLogicDS::performStartSync(void)
     sta=implEndDataRead();
     // show items
     PDEBUGPRINTFX(DBG_HOT,("%s: number of local items involved in %ssync = %ld",getName(), fSlowSync ? "slow " : "",(long)fItems.size()));
-    CONSOLEPRINTF(("  %ld local items are new/changed/deleted for this sync",fItems.size()));
+    CONSOLEPRINTF(("  %ld local items are new/changed/deleted for this sync",(long)fItems.size()));
     if (PDEBUGTEST(DBG_DATA+DBG_DETAILS)) {
       PDEBUGBLOCKFMTCOLL(("SyncSet","Items involved in Sync","datastore=%s",getName()));
       for (TSyncItemPContainer::iterator pos=fItems.begin();pos!=fItems.end();pos++) {
